CCustomSscript class moved to separate files.
