+++ CLEO5.vcxproj
@@ -128,6 +128,7 @@
     <ClInclude Include="cleo_sdk\CLEO_Utils.h" />
     <ClInclude Include="source\CCodeInjector.h" />
     <ClInclude Include="source\CCustomOpcodeSystem.h" />
+    <ClInclude Include="source\CCustomScript.h" />
     <ClInclude Include="source\CDebug.h" />
     <ClInclude Include="source\CDmaFix.h" />
     <ClInclude Include="source\CGameMenu.h" />
