+++ source/CCustomScript.cpp
@@ -1,0 +1,498 @@
+#include "stdafx.h"
+#include "CCustomScript.h"
+#include "CleoBase.h"
+#include "crc32.h"
+
+
+using namespace std;
+using namespace CLEO;
+
+CCustomScript::CCustomScript(const char* szFileName, bool bIsMiss, CRunningScript* parent, int label) : CRunningScript()
+{
+    TRACE(""); // separator
+    TRACE("Loading custom script '%s'...", szFileName);
+
+    bIsCustom = true;
+    bIsMission = bUseMissionCleanup = bIsMiss;
+
+    try
+    {
+        ifstream is;
+        if (label != 0) // Create external from label.
+        {
+            if (!parent)
+                throw logic_error("Trying to create external thread from label without parent thread");
+
+            if (!parent->IsCustom())
+                throw logic_error("Only custom threads can spawn children threads from label");
+
+            auto cs = (CCustomScript*)parent;
+
+            m_compatVer = cs->GetCompatibility();
+            m_debugMode = cs->GetDebugMode();
+            m_scriptFileDir = cs->GetScriptFileDir();
+            m_scriptFileName = cs->GetScriptFileName();
+            m_workDir = cs->GetWorkDir();
+
+            BaseIP = cs->GetBasePointer();
+            CurrentIP = cs->GetBasePointer() - label;
+            memcpy(Name, cs->Name, sizeof(Name));
+            m_codeChecksum = cs->m_codeChecksum;
+            m_parentThread = cs;
+            cs->m_childThreads.push_back(this);
+        }
+        else
+        {
+            // store script file directory and name
+            string pathStr = szFileName;
+            FilepathNormalize(pathStr, false);
+            FS::path path = pathStr;
+
+            // file exists?
+            if (!FS::is_regular_file(path))
+            {
+                if (path.extension() == cs_ext)
+                {
+                    // maybe it was renamed to enable compatibility mode?
+                    auto compatPath = path;
+
+                    compatPath.replace_extension(cs4_ext);
+                    if (FS::is_regular_file(compatPath))
+                    {
+                        path = compatPath;
+                    }
+                    else
+                    {
+                        compatPath.replace_extension(cs3_ext);
+                        if (FS::is_regular_file(compatPath))
+                        {
+                            path = compatPath;
+                        }
+                        else
+                        {
+                            throw logic_error("File does not exists");
+                        }
+                    }
+                }
+                else
+                    throw logic_error("File does not exists");
+            }
+
+            // deduce compatibility mode from filetype extension
+            if (path.extension() == cs4_ext)
+                m_compatVer = CLEO_VER_4;
+            else
+                if (path.extension() == cs3_ext)
+                    m_compatVer = CLEO_VER_3;
+
+            if (m_compatVer == CLEO_VER_CUR && parent != nullptr)
+            {
+                // inherit compatibility mode from parent
+                m_compatVer = CLEO_GetScriptVersion(parent);
+
+                // try loading file with same compatibility mode filetype extension
+                auto compatPath = path;
+                if (m_compatVer == CLEO_VER_4)
+                {
+                    compatPath.replace_extension(cs4_ext);
+                    if (FS::is_regular_file(compatPath))
+                        path = compatPath;
+                }
+                else
+                    if (m_compatVer == CLEO_VER_3)
+                    {
+                        compatPath.replace_extension(cs3_ext);
+                        if (FS::is_regular_file(compatPath))
+                            path = compatPath;
+                    }
+            }
+
+            m_scriptFileDir = path.parent_path().string();
+            m_scriptFileName = path.filename().string();
+
+            if (parent != nullptr)
+            {
+                m_debugMode = ((CCustomScript*)parent)->GetDebugMode();
+                m_workDir = ((CCustomScript*)parent)->GetWorkDir();
+            }
+            else
+            {
+                m_debugMode = CleoInstance.ScriptEngine.NativeScriptsDebugMode; // global setting
+                m_workDir = Filepath_Game; // game root
+            }
+
+            ifstream is(path.string().c_str(), ios::binary);
+            is.exceptions(ios::badbit | ios::failbit);
+            size_t length;
+            is.seekg(0, ios::end);
+            length = (size_t)is.tellg();
+            is.seekg(0, ios::beg);
+
+            if (bIsMiss)
+            {
+                if (CTheScripts::bAlreadyRunningAMissionScript)
+                    throw logic_error("Starting of custom mission when other mission loaded");
+
+                CTheScripts::bAlreadyRunningAMissionScript = true;
+                MissionIndex = -1;
+                BaseIP = CurrentIP = missionBlock; // TODO: there should be check length <= missionBlock size
+            }
+            else
+            {
+                BaseIP = CurrentIP = new BYTE[length];
+            }
+            is.read(reinterpret_cast<char*>(BaseIP), length);
+
+            m_codeSize = length;
+            m_codeChecksum = crc32(reinterpret_cast<BYTE*>(BaseIP), length);
+
+            // thread name from filename
+            auto threadNamePath = path;
+            if (threadNamePath.extension() == cs3_ext || threadNamePath.extension() == cs4_ext)
+            {
+                threadNamePath.replace_extension(cs_ext); // keep original extension even in compatibility modes
+            }
+            auto fName = threadNamePath.filename().string();
+
+            memset(Name, '\0', sizeof(Name));
+            if (!fName.empty())
+            {
+                auto len = min(fName.length(), sizeof(Name) - 1); // and text terminator
+                memcpy(Name, fName.c_str(), len);
+            }
+        }
+        CleoInstance.ScriptEngine.LastScriptCreated = this;
+        m_ok = true;
+    }
+    catch (exception& e)
+    {
+        LOG_WARNING(0, "Error during loading of custom script '%s' occured.\nError message: %s", szFileName, e.what());
+    }
+    catch (...)
+    {
+        LOG_WARNING(0, "Unknown error during loading of custom script '%s' occured.", szFileName);
+    }
+}
+
+CCustomScript::~CCustomScript()
+{
+    if (BaseIP && !bIsMission) delete[] BaseIP;
+    CleoInstance.OpcodeSystem.scriptDeleteDelegate(this);
+
+    if (CleoInstance.ScriptEngine.LastScriptCreated == this) CleoInstance.ScriptEngine.LastScriptCreated = nullptr;
+
+    // TODO: release m_scriptSprites
+}
+
+void CCustomScript::AddScriptToList(CRunningScript** queuelist)
+{
+    ((::CRunningScript*)this)->AddScriptToList((::CRunningScript**)queuelist); // CRunningScript from Plugin SDK
+}
+
+void CCustomScript::RemoveScriptFromList(CRunningScript** queuelist)
+{
+    ((::CRunningScript*)this)->RemoveScriptFromList((::CRunningScript**)queuelist); // CRunningScript from Plugin SDK
+}
+
+void CCustomScript::Process()
+{
+    // save native scripts' drawing state
+    CleoInstance.ScriptEngine.StoreScriptDrawsState();
+
+    // initialize script drawings
+    m_drawingNow = true;
+    CTheScripts::UseTextCommands = m_useTextCommands;
+    CTheScripts::NumberOfIntroRectanglesThisFrame = 0;
+    memcpy(CTheScripts::ScriptSprites, m_scriptSprites.data(), sizeof(m_scriptSprites)); // restore loaded textures
+    CTheScripts::NumberOfIntroTextLinesThisFrame = 0;
+    CScriptEngine::SetScriptTextsDefaults();
+
+    ProcessScript(this);
+
+    // store this script's drawings
+    m_drawingNow = false;
+    m_useTextCommands = CTheScripts::UseTextCommands;
+    if (m_useTextCommands)
+    {
+        m_scriptDraws.assign(
+            CTheScripts::IntroRectangles,
+            CTheScripts::IntroRectangles + min(CTheScripts::NumberOfIntroRectanglesThisFrame, CScriptEngine::Script_Draws_Capacity) // clamp, as original SA scripts seems do not care
+        );
+
+        m_scriptTexts.assign(
+            CTheScripts::IntroTextLines,
+            CTheScripts::IntroTextLines + min(CTheScripts::NumberOfIntroTextLinesThisFrame, CScriptEngine::Script_Texts_Capacity)
+        );
+    }
+    else
+    {
+        m_scriptDraws.clear();
+        m_scriptTexts.clear();
+    }
+    memcpy(m_scriptSprites.data(), CTheScripts::ScriptSprites, sizeof(m_scriptSprites)); // store loaded textures
+
+    // restore native scripts' drawing state
+    CleoInstance.ScriptEngine.RestoreScriptDrawsState();
+}
+
+void CCustomScript::ShutdownThisScript()
+{
+    ((::CRunningScript*)this)->ShutdownThisScript(); // CRunningScript from Plugin SDK
+}
+
+void CCustomScript::Draw(uchar beforeFade)
+{
+    if (m_useTextCommands == false ||
+        (m_scriptDraws.empty() && m_scriptTexts.empty()))
+    {
+        return; // nothing to draw
+    }
+
+    // save native scripts' drawing state
+    CleoInstance.ScriptEngine.StoreScriptDrawsState();
+
+    // apply this script's draws
+    m_drawingNow = true;
+    CTheScripts::UseTextCommands = m_useTextCommands;
+
+    CTheScripts::NumberOfIntroRectanglesThisFrame = (WORD)m_scriptDraws.size();
+    memcpy(CTheScripts::IntroRectangles, m_scriptDraws.data(), m_scriptDraws.size() * sizeof(decltype(m_scriptDraws)::value_type));
+
+    memcpy(CTheScripts::ScriptSprites, m_scriptSprites.data(), sizeof(m_scriptSprites));
+
+    CTheScripts::NumberOfIntroTextLinesThisFrame = (WORD)m_scriptTexts.size();
+    memcpy(CTheScripts::IntroTextLines, m_scriptTexts.data(), m_scriptTexts.size() * sizeof(decltype(m_scriptTexts)::value_type));
+
+    // draw
+    if (beforeFade) CleoInstance.ScriptEngine.DrawScriptTextBeforeFade_Orig(beforeFade);
+    else CleoInstance.ScriptEngine.DrawScriptTextAfterFade_Orig(beforeFade);
+
+    // restore native scripts' drawing state
+    m_drawingNow = false;
+    CleoInstance.ScriptEngine.RestoreScriptDrawsState();
+}
+
+CSprite2d* CCustomScript::GetScriptSprite(size_t index)
+{
+    if (index == 0 || index > CScriptEngine::Script_Sprites_Capacity)
+    {
+        return nullptr; // invalid argument
+    }
+
+    index -= 1; // 1-based to 0-based
+
+    if (index == 0 || index >= m_scriptSprites.size())
+    {
+        return nullptr;
+    }
+
+    return m_drawingNow ? &CTheScripts::ScriptSprites[index] : &m_scriptSprites[index];
+}
+
+bool CCustomScript::GetDebugMode() const
+{
+    if (!bIsCustom)
+        return CleoInstance.ScriptEngine.NativeScriptsDebugMode;
+
+    return m_debugMode;
+}
+
+void CCustomScript::SetDebugMode(bool enabled)
+{
+    if (!bIsCustom)
+        CleoInstance.ScriptEngine.NativeScriptsDebugMode = enabled;
+    else
+        m_debugMode = enabled;
+}
+
+const char* CCustomScript::GetScriptFileDir() const
+{
+    if (!bIsCustom)
+        return CleoInstance.ScriptEngine.MainScriptFileDir.c_str();
+
+    return m_scriptFileDir.c_str();
+}
+
+void CCustomScript::SetScriptFileDir(const char* directory)
+{
+    if (!bIsCustom)
+        CleoInstance.ScriptEngine.MainScriptFileDir = directory;
+    else
+        m_scriptFileDir = directory;
+}
+
+const char* CCustomScript::GetScriptFileName() const
+{
+    if (!bIsCustom)
+        return CleoInstance.ScriptEngine.MainScriptFileName.c_str();
+
+    return m_scriptFileName.c_str();
+}
+
+void CCustomScript::SetScriptFileName(const char* filename)
+{
+    if (!bIsCustom)
+        CleoInstance.ScriptEngine.MainScriptFileName = filename;
+    else
+        m_scriptFileName = filename;
+}
+
+string CCustomScript::GetScriptFileFullPath() const
+{
+    string path = GetScriptFileDir();
+    path += '\\';
+    path += GetScriptFileName();
+    return path;
+}
+
+const char* CCustomScript::GetWorkDir() const
+{
+    if (!bIsCustom)
+        return CleoInstance.ScriptEngine.MainScriptCurWorkDir.c_str();
+
+    return m_workDir.c_str();
+}
+
+void CCustomScript::SetWorkDir(const char* directory)
+{
+    if (directory == nullptr || strlen(directory) == 0)
+        return;  // Already done. Empty path is relative path starting at current work dir
+
+    auto resolved = ResolvePath(directory);
+
+    if (!bIsCustom)
+        CleoInstance.ScriptEngine.MainScriptCurWorkDir = resolved;
+    else
+        m_workDir = resolved;
+}
+
+string CCustomScript::ResolvePath(const char* path, const char* customWorkDir) const
+{
+    if (path == nullptr)
+    {
+        return {};
+    }
+
+    auto fsPath = FS::path(path);
+
+    // check for virtual path root
+    enum class VPref { None, Game, User, Script, Cleo, Modules } virtualPrefix = VPref::None;
+    if (!fsPath.empty())
+    {
+        const auto root = fsPath.begin()->string(); // first path element
+        const auto r = root.c_str();
+
+        if (_strcmpi(r, DIR_GAME) == 0) virtualPrefix = VPref::Game;
+        else if (_strcmpi(r, DIR_USER) == 0) virtualPrefix = VPref::User;
+        else if (_strcmpi(r, DIR_SCRIPT) == 0 && !IsLegacyScript((CRunningScript*)this)) virtualPrefix = VPref::Script;
+        else if (_strcmpi(r, DIR_CLEO) == 0) virtualPrefix = VPref::Cleo;
+        else if (_strcmpi(r, DIR_MODULES) == 0) virtualPrefix = VPref::Modules;
+    }
+
+    // not virtual
+    if (virtualPrefix == VPref::None)
+    {
+        if (fsPath.is_relative())
+        {
+            if (customWorkDir != nullptr)
+                fsPath = ResolvePath(customWorkDir) / fsPath;
+            else
+                fsPath = GetWorkDir() / fsPath;
+        }
+
+        auto result = fsPath.string();
+        FilepathNormalize(result, false);
+
+        // ModLoader support: make paths withing game directory relative to it
+        FilepathRemoveParent(result, Filepath_Game);
+
+        return move(result);
+    }
+
+    // expand virtual paths
+    FS::path resolved;
+    switch (virtualPrefix)
+    {
+        case VPref::User: resolved = Filepath_User; break;
+        case VPref::Script: resolved = GetScriptFileDir(); break;
+        case VPref::Game: resolved = Filepath_Game; break;
+        case VPref::Cleo: resolved = Filepath_Cleo; break;
+        case VPref::Modules: resolved = Filepath_Cleo + "\\cleo_modules"; break;
+        default: resolved = "<error>"; break; // should never happen
+    }
+
+    // append all but virtual prefix from original path
+    for (auto it = ++fsPath.begin(); it != fsPath.end(); it++)
+        resolved /= *it;
+
+    auto result = resolved.string();
+    FilepathNormalize(result, false);
+    return move(result);
+}
+
+string CCustomScript::GetInfoStr(bool currLineInfo) const
+{
+    ostringstream ss;
+
+    auto threadName = GetName();
+    auto fileName = GetScriptFileName();
+
+    if (memcmp(threadName.c_str(), fileName, threadName.length()) != 0) // thread name no longer same as filename (was set with 03A4)
+    {
+        ss << "'" << threadName << "' from ";
+    }
+
+    ss << "'" << fileName << "'";
+
+    if (currLineInfo)
+    {
+        ss << " at ";
+
+        if (false)
+        {
+            // TODO: get Sanny's SMC extra info
+            ss << "line " << 0;
+            ss << " - ";
+            ss << "CODE";
+        }
+        else
+        {
+            auto base = (DWORD)BaseIP;
+            if (base == 0) base = (DWORD)scmBlock;
+            auto currPos = (DWORD)CCustomOpcodeSystem::lastOpcodePtr;
+
+            if (IsMission() && !IsCustom())
+            {
+                if (currPos >= (DWORD)missionBlock)
+                {
+                    // we are in mission code buffer
+                    // native missions are loaded from script file into mission block area
+                    currPos += ((DWORD*)CTheScripts::MultiScriptArray)[MissionIndex]; // start offset of this mission within source script file
+                }
+                else
+                {
+                    base = (DWORD)scmBlock; // seems that mission uses main scm code
+                }
+            }
+
+            auto offset = currPos - base;
+
+            ss << "offset {" << offset << "}"; // Sanny offsets style
+            ss << " - ";
+            ss << std::hex << std::uppercase << std::setw(4) << std::setfill('0') << CCustomOpcodeSystem::lastOpcode;
+
+            auto commandName = CleoInstance.OpcodeInfoDb.GetCommandName(CCustomOpcodeSystem::lastOpcode);
+            if (commandName != nullptr)
+            {
+                ss << ": " << commandName;
+            }
+            else
+            {
+                ss << ": ...";
+            }
+        }
+    }
+
+    return ss.str();
+}
+
