+++ source/exports.cpp
@@ -1,5 +1,6 @@
 #include "stdafx.h"
 #include "CleoBase.h"
+#include "CCustomScript.h"
 
 
 namespace CLEO
@@ -413,12 +414,14 @@ extern "C"
 
     DWORD WINAPI CLEO_GetScriptTextureById(CLEO::CRunningScript* thread, int id)
     {
-        CCustomScript* customScript = reinterpret_cast<CCustomScript*>(thread);
-        // We need to store-restore to update the texture list, not optimized, but this will not be used every frame anyway
-        customScript->StoreScriptDraws();
-        auto* texture = customScript->GetScriptSprite(id - 1);
-        customScript->ApplyScriptDraws();
-        return (DWORD)texture;
+        if (thread->IsCustom())
+        {
+            return (DWORD)reinterpret_cast<CCustomScript*>(thread)->GetScriptSprite(id);
+        }
+        else
+        {
+            return (DWORD)nullptr; // TODO: handle properly for native scripts
+        }
     }
 
     DWORD WINAPI CLEO_GetInternalAudioStream(CLEO::CRunningScript* unused, DWORD audioStreamPtr)
